make zap functions hook, tidy up ui
