+++ src/components/Row/index.tsx
@@ -12,10 +12,9 @@ const Row = styled(Box)<{
 }>`
   width: ${({ width }) => width ?? '100%'};
   display: flex;
-  padding: 0;
   align-items: ${({ align }) => align ?? 'center'};
   justify-content: ${({ justify }) => justify ?? 'flex-start'};
-  padding: ${({ padding }) => padding};
+  padding: ${({ padding }) => padding ?? 0};
   border: ${({ border }) => border};
   border-radius: ${({ borderRadius }) => borderRadius};
   gap: ${({ gap }) => gap};
